app/services/caption_service.py
from transformers import BlipProcessor, BlipForConditionalGeneration
from PIL import Image

processor = None
model = None

def _load_models():
    global processor, model
    if processor is None or model is None:
        processor = BlipProcessor.from_pretrained(
            "Salesforce/blip-image-captioning-base"
        )
        model = BlipForConditionalGeneration.from_pretrained(
            "Salesforce/blip-image-captioning-base"
        )

def generate_caption(image_path: str):
    _load_models()
    
    image = Image.open(image_path).convert("RGB")
    inputs = processor(image, return_tensors="pt")
    output = model.generate(**inputs)
    caption = processor.decode(output[0], skip_special_tokens=True)

    return caption
