requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
torch
transformers
ultralytics
